learning_journey/Day5_ReAct_muti_agent/requirements.txt
langchain
requests
feedparser
python-dotenv
openai   # 用于调用DeepSeek接口
